LPS-12802 Dynamic data source support for counterDataSource
